move account shared components
