+++ dashboard/src/components/accounts/Accounts.vue
@@ -46,7 +46,7 @@
 <script lang="ts">
 import { Component, Prop, Vue, Watch } from 'vue-property-decorator';
 import keyring from '@polkadot/ui-keyring';
-import Keypair from '../shared/Keypair.vue';
+import Keypair from '../shared/accounts/Keypair.vue';
 import { u8aToHex } from '@polkadot/util';
 import Connector from '@vue-polkadot/vue-api';
