+++ dashboard/src/components/shared/accounts/Keypair.vue
