+++ dashboard/src/views/AccountsBackup.vue
@@ -12,7 +12,7 @@
 </template>
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import Backup from '../components/shared/Backup.vue';
+import Backup from '../components/shared/accounts/Backup.vue';
 import ModalView from '@/components/shared/modals/ModalView.vue';
 
 @Component({
