+++ dashboard/src/views/AccountsChangePassword.vue
@@ -10,7 +10,7 @@
 </template>
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import Changepass from '../components/shared/ChangePassword.vue';
+import Changepass from '../components/shared/accounts/ChangePassword.vue';
 import ModalView from '@/components/shared/modals/ModalView.vue';
 @Component({
   components: {
