+++ dashboard/src/views/AccountsCreate.vue
@@ -11,7 +11,7 @@
 </template>
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import Create from '../components/shared/Create.vue';
+import Create from '../components/shared/accounts/Create.vue';
 import ModalView from '@/components/shared/modals/ModalView.vue';
 
 @Component({
