+++ dashboard/src/views/AccountsRestore.vue
@@ -9,7 +9,7 @@
 </template>
 <script lang="ts">
 import { Component, Prop, Vue } from 'vue-property-decorator';
-import Restore from '../components/shared/Restore.vue';
+import Restore from '../components/shared/accounts/Restore.vue';
 import ModalView from '@/components/shared/modals/ModalView.vue';
 
 @Component({
